% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_entities.R
\name{add_entities}
\alias{add_entities}
\alias{print.survey_ents}
\title{Add a summary of entities to a survey}
\usage{
add_entities(srv, entities = NULL)

\method{print}{survey_ents}(ents, width = getOption("width"))
}
\arguments{
\item{srv}{A survey object.}

\item{entity}{Optional: Specify a \code{data.frame} which contains the summary
and information for entities.}
}
\description{
Third step in creating a \code{survey} that can be used with \code{generate_report}.
The function accepts a \code{data.frame} as input, but columns must be correctly named
or be dropped (see description below). If no data is provided \code{(entities = NULL)},
\code{add_entities} will generate a suggestion based on the data.
}
\note{
This function requires that mainentity is specified in the measurement model.
See \code{set_association} for more information.
}
\section{Structure/naming convention for \code{ents}}{


\describe{

   \item{\code{entity}}{List of names for \code{mainentity} found in the data.}

   \item{\code{n}}{The number of rows/observations per entity.}

   \item{\code{marketshare}}{The marketshare associated with each entity.
   This is used by \code{add_weigts} when calculating weights.}

   \item{\code{other}}{Specifies whether the entity should be collected among
   "other" entities. I.e., when it is too small to publish and should be
   excluded from reports.}

}

A structure and naming convention is also required for the survey itself, see
\code{help(survey) for information.}
}
\examples{
x <- survey(data.frame("test" = 1, stringsAsFactors = FALSE))
x \%>\% add_mm() \%>\% add_entities()
}
\author{
Kristian D. Olsen
}

