% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_mm.R
\name{add_mm}
\alias{add_mm}
\alias{get_question}
\alias{print.survey_mm}
\title{Add measurement model to a survey}
\usage{
add_mm(srv, mm = NULL)

get_question(srv, vars)

\method{print}{survey_mm}(mm, width = getOption("width"))
}
\arguments{
\item{srv}{A survey object.}

\item{mm}{Optional: Specify a \code{data.frame} which contains the measurement
model.}
}
\description{
Second step in creating a \code{survey} that can be used with \code{generate_report}.
The function accepts a \code{data.frame} as input, but columns must be correctly named
or be dropped (see description below). If no model is provided \code{(mm = NULL)},
\code{add_mm} will generate a suggestion based on the data.
}
\note{
For smaller surveys, using this function without specifiying a model might
not show all possible values for factor variables etc.
}
\section{Structure/naming convention for \code{mm}}{


\describe{

   \item{\code{latent}}{A column describing special significance of variables,
   such as latent association, complaints and so forth.}

   \item{\code{manifest}}{The name of variables/columns in the data.}

   \item{\code{question}}{Questions as posed during interviews. Used
   for printed values in \code{generate_report}.}

   \item{\code{type}}{The type of variable in the data. Either \code{character},
   \code{factor}, \code{numeric}, or \code{scale}. If type is specified as factor,
   the values (see below) will be used to create ordered factors.}

   \item{\code{values}}{Optional: Possible values (factor levels) for variables
   in the data. If \code{type} is \code{scale}, values are the outer points (1 and 10),
   and an optional "do not know".}

}

A structure and naming convention is also required for the survey itself, see
\code{help(survey) for information.}
}
\examples{
x <- survey(data.frame("test" = 1, stringsAsFactors = FALSE))
x \%>\% add_mm()
}
\author{
Kristian D. Olsen
}

