% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_weights.R
\name{add_weights}
\alias{add_weights}
\alias{add_weights.character}
\alias{add_weights.factor}
\alias{add_weights.survey}
\title{Add weights based on marketshares}
\usage{
add_weights(x, entities = NULL)

\method{add_weights}{survey}(srv)

\method{add_weights}{character}(x, ms)

\method{add_weights}{factor}(x, ms)
}
\arguments{
\item{x}{A survey object, or vector of type \code{character} or \code{facotor}.}

\item{entities}{Optional: If a vector is passed to \code{x}, the entities/ents
from a survey object must be passed along with it.}
}
\description{
This function calculates weights for each entity based on their expected
marketshares and number of observations. If a \code{survey} is passed to the function,
a new column \code{w} is added to the \code{df} which contains a weight for all
the obserations. If a vector of types \code{character} or \code{factor} is
passed to the function, and the entities/ents from a survey, a \code{numeric}
vector is returned instead.
}
\note{
See \code{add_entities} for information on the entities object.
}
\examples{
x <- survey(data.frame("test" = 1, stringsAsFactors = FALSE))
x <- x \%>\% add_mm() \%>\% add_entities() \%>\% add_weights()
x \%>\% mutate(w = add_weights(q1, x$ents)) # i.e. create weights for subsets
}
\author{
Kristian D. Olsen
}

