% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survey.R
\name{as.survey}
\alias{as.survey}
\alias{is.survey}
\alias{print.survey}
\title{Create a survey object}
\usage{
as.survey(x)

is.survey(x)

\method{print}{survey}(lst, width = getOption("width"))
}
\arguments{
\item{x}{The list to convert to a survey}
}
\description{
This function takes a list as input, replaces names (data = df etc), and checks
whether the necessary info is in place to generate a report. If this is the case,
the class 'survey' is appended to the object.
}
\note{
survey objects have a print method.
}
\examples{
survey <- as.survey(lst)
}
\author{
Kristian D. Olsen
}

