% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survey.R
\name{as.survey_ents}
\alias{as.survey_ents}
\alias{is.survey_ents}
\alias{print.survey_ents}
\title{Create a survey_ents object}
\usage{
as.survey_ents(x)

is.survey_ents(x)

\method{print}{survey_ents}(ents, width = getOption("width"))
}
\arguments{
\item{x}{A data.frame}
}
\description{
Convert a data.frame to a survey_ents object if the required data is present.
}
\note{
survey_ents objects have a print method.
}
\examples{
survey$ents <- as.survey_ents(survey$ents)
}
\author{
Kristian D. Olsen
}

