% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beamer_report.R
\name{beamer_template}
\alias{beamer_template}
\alias{copy_beamer_theme}
\alias{generate_beamer}
\title{Beamer template}
\usage{
beamer_template(dev = "cairo_pdf", toc = TRUE, keep_tex = FALSE,
  slide_level = NULL)

copy_beamer_theme(dir = NULL)

generate_beamer(entity, md, dir, envir)
}
\arguments{
\item{dev}{Defaults to cairo_pdf. Set to 'pdf' if you do not have the
'cairo_pdf' device.}

\item{toc}{Include a table of contents or not. Default is true.}

\item{keep_tex}{Keep the intermediate .tex files (useful for troubleshooting)}

\item{slide_level}{Manually set the slide level. Default is 2, which makes
a single hashtag a section divider in latex and double hashtag into frames.}
}
\description{
This is a beamer template for generating reports, and it is based on
mtheme. The template assumes that .sty files are in the same directory as
the \code{.Rmd} file that is being rendered to a beamer pdf.
}
\note{
This function uses a slightly modified template_beamer.tex file which
is included in the package (i.e. does not rely on rmarkdown).
}
\examples{
render("index.Rmd", "reporttool::report_template")
}
\author{
Kristian D. Olsen
}

