% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_date.R
\name{convert_date}
\alias{convert_date}
\title{Convert columns to date}
\usage{
convert_date(survey, ...)
}
\arguments{
\item{survey}{A survey object.}

\item{...}{Columns to convert to dates. Of the format \code{column_name = date_format}.}
}
\description{
This function lets you convert columns with dates to \code{Date}.
}
\examples{
x \%>\% convert_date("startdate" = "\%m/\%d/\%Y \%H:\%M:\%S")
}
\author{
Kristian D. Olsen
}

