% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_factor.R
\name{convert_factor}
\alias{convert_factor}
\title{Convert columns to factor}
\usage{
convert_factor(survey, ...)
}
\arguments{
\item{survey}{A survey object.}

\item{...}{Columns to convert to factors. Of the format \code{column_name = factor_levels}.}
}
\description{
This function lets you convert columns to \code{factor}.
}
\examples{
x \%>\% convert_factor("q1" = "reporttool")
}
\author{
Kristian D. Olsen
}

