% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{evaluate_rmd}
\alias{evaluate_rmd}
\title{Evaluate a .Rmd file (after converting with rmd_to_r)}
\usage{
evaluate_rmd(rmd, envir = parent.frame())
}
\arguments{
\item{rmd}{Path to a rmarkdown file.}

\item{envir}{The environment in which to evaluate the rmarkdown file.}
}
\description{
Evaluate a \code{.Rmd} file in a given environment and return the results,
including text with inline code (e.g. \code{`r 1+1`} becomes 2).
}
\examples{
evaluate_rmd("test.Rmd", envir = parent.frame())
}
\author{
Kristian D. Olsen
}

