% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factor_data.R
\name{factor_data}
\alias{factor_data}
\title{Convert variables to factor}
\usage{
factor_data(survey, vars = NULL)
}
\arguments{
\item{survey}{A survey object.}

\item{vars}{The variables to convert. If \code{NULL}, variables specified as
\code{factor} in the measurement model will be converted. Specify if you want
to convert \code{scale} variables.}
}
\description{
A convenience function to convert variables in the data to (ordered) factor variables,
based on the type and values columns in the measurement model.
}
\note{
See \code{rescale_score} and \code{clean_score} to convert scales from
factor to numeric again.
}
\examples{
x \%>\% factor_data(vars = c("q1", "q17"))
}
\author{
Kristian D. Olsen
}

