% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowchart.R
\name{flowchart}
\alias{flowchart}
\title{Flowchart}
\usage{
flowchart(...)
}
\arguments{
\item{scores}{The scores/results for each latent.}

\item{weights}{A vector with inner weights from PLS-PM.}

\item{latents}{The names to used for the aspects in the EPSI model. (language)}
}
\description{
Create a flowchart with weights from PLS-PM (inner weights).
}
\note{
Alternately you can supply a lone survey object.
}
\author{
Kristian D. Olsen
}

