% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_input.R
\name{from_clipboard}
\alias{from_clipboard}
\title{Read from Windows/OSX clipboard}
\usage{
from_clipboard(header = TRUE, sep = "", dec = ".", encoding = "")
}
\arguments{
\item{sep}{The}

\item{encoding}{The encoding to use when writing.}
}
\description{
Thin wrapper for reading from windows/OSX clipboards with the most-used defaults.
}
\note{
This function only works on Windows or OSX, and the data-size cannot
exceed 128kb in Windows.
}
\examples{
x <- from_clipboard()
}
\author{
Kristian D. Olsen
}

