% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_input.R
\name{from_directory}
\alias{from_directory}
\title{Extract from a study directory on sharepoint}
\usage{
from_directory(file)
}
\arguments{
\item{file}{Path to a study directory on the intranet.}
}
\value{
A list containing the EM-data, entities (observations, marketshare),
and a measurement model (latent association, question text, values etc.)
}
\description{
This function extracts whatever data is available in our standard file
structure and tries to structure it like a input.xlsx file.
}
\examples{
x <- from_directory("https://the.intranet.se/EPSI/example")
}
\author{
Kristian D. Olsen
}

