% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report_ppt.R
\name{generate_ppt}
\alias{generate_ppt}
\title{Powerpoint template}
\usage{
generate_ppt(entity, dir, envir)
}
\arguments{
\item{entity}{The name of the entity to create a report for.}

\item{dir}{The directory in which to place the PPT folder containing the reports.}

\item{envir}{The environment in which to evaluate the rmarkdown code.}

\item{rmd}{A loaded (\code{readLines}) rmarkdown document to use as a template.}
}
\description{
This function uses \code{ReporteRs} and \code{evaluate} to generate powerpoint
slides based on the same rmarkdown syntax that creates the pdf reports. I.e.
it is a dumbed down version of normal rmarkdown, where only \code{#} and \code{##}
are understood as being section titles and new frames/subtitles respectively.
}
\note{
This function is used by \code{generate_report}.
}
\examples{
x <- survey(data.frame("test" = 1, stringsAsFactors = FALSE))
x \%>\% add_mm()
}
\author{
Kristian D. Olsen
}

