% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate report}
\usage{
generate_report(report = NULL, entity = NULL, data = NULL, type = "pdf")
}
\arguments{
\item{report}{Path to the report template (.Rmd) for a specific study.}

\item{entity}{Optional: Name of a specific entity to create a report for.}

\item{data}{Optional: path to input.xlsx (if NULL, assumes that this file can
be found in working directory.)}
}
\value{
Nothing. A .Rmd file and desired output is created for each entity in
the data, or as specified.
}
\description{
Generate one or more reports using the templates in this package.
}
\examples{
generate_report("/Internal/Internal_report_2014.Rmd", entity=c("EPSI", "SKI"))
}
\author{
Kristian D. Olsen
}

