% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_report.R
\name{get_survey}
\alias{get_survey}
\title{Get survey data}
\usage{
get_survey(file)
}
\arguments{
\item{file}{Path to the input file}
}
\value{
Either a list with the
}
\description{
This function reads in and converts data to a survey. Obviously requires that
whatever data is read in can be converted to a survey, which means that the file-type
will usually be .xlsx or .Rdata.
}
\examples{
generate_report("/Internal/Internal_report_2014.Rmd", entity=c("EPSI", "SKI"))
}
\author{
Kristian D. Olsen
}

