% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report_ioslides.R
\name{io_template}
\alias{io_template}
\title{ioslides template}
\usage{
io_template(dev = "svg", smaller = TRUE)
}
\arguments{
\item{dev}{Defaults to svg for higher quality graphics. Standard in ioslides
is png.}

\item{smaller}{Use smaller text for the slides by default.}
}
\description{
Template for generating ioslides reports.
}
\note{
Simple wrapper. Custom CSS is on the way.
}
\examples{
render("index.Rmd", "reporttool::io_template")
}
\author{
Kristian D. Olsen
}

