% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare survey data}
\usage{
prepare_data(survey, type = "mean", cutoff = 0.3)
}
\arguments{
\item{survey}{A survey object.}

\item{type}{The approach to use for latents. Either \code{mean}, \code{pls},
\code{pls-wizard} (no latents or EM-variables) or \code{none} (no latents).}
}
\value{
Returns the survey with EM-variables and latent scores added using the
specified method.
}
\description{
This function does the imputation/PLS-PM (or simple mean) for the survey data.
In addition, it calculates missing percentage and tallies the valid observations
for each entity.
}
\author{
Kristian D. Olsen
}

