% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_report.R
\name{prepare_report}
\alias{prepare_report}
\title{Prepare survey for reporting}
\usage{
prepare_report(srv)
}
\arguments{
\item{srv}{A survey object.}
}
\description{
This function prepares a survey for reporting.
}
\examples{
x \%>\% prepare_report()
}
\author{
Kristian D. Olsen
}

