% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_csv}
\alias{read_csv}
\title{Collate one or more files to a list (for an EPSI-object).}
\usage{
read_csv(file, encoding)
}
\arguments{
\item{encoding}{Encoding for files (if reading directory or csv). Defaults
to 'latin1'.}

\item{input}{file to a directory or xlsx-file with the input data}

\item{contrast}{Optional file to historic data (xlsx or csv format)}

\item{historic}{Optional file to historic data (xlsx or csv format)}
}
\value{
A list of data.frames from xlsx sheets and/or other files.
}
\description{
This function takes one or more files (or a directory) and collates them to
a list to be converted to a EPSI-object. It is not meant for other
uses. The function only looks for prespecified structures.
}
\note{
This function requires openxlsx. The function attempts to read
CSV-files as semicolon separated first, and then comma.
}
\examples{
x <- get_input("input.xlsx", "contrast.csv")
}
\author{
Kristian D. Olsen
}
\seealso{
\code{\link{read_sheets}} for a function that reads all sheets
of a xlsx file to a list.
}

