% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_dir}
\alias{read_dir}
\title{Read common data formats}
\usage{
read_dir(file)
}
\arguments{
\item{file}{Path to a study directory on the intranet.}
}
\value{
A list containing the EM-data, entities (observations, marketshare),
and a measurement model (latent association, question text, values etc.)
}
\description{
A simple wrapper for reading data. Currently supports txt,
csv, csv2 and xlsx. The function also lowercases all list and column names,
and cleans common strings for missing values.
}
\examples{
x <- read_dir("https://the.intranet.se/EPSI/example")
}
\author{
Kristian D. Olsen
}

