% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_sheets}
\alias{read_sheets}
\title{Read sheets from xlsx-files to a list.}
\usage{
read_sheets(file, sheets = NULL, clean.missing = FALSE)
}
\arguments{
\item{file}{The path to the .xlsx file.}

\item{sheets}{Optional: Specify the sheets to be read.}

\item{clean.na}{Set to TRUE to clean typical NA-strings from .xlsx file.}
}
\value{
A list containing data.frames matching the sheets in the .xlsx file.
If only one sheet is read, the function returns a data.frame instead.
}
\description{
This function (a very thin wrapper for openxlsx functions) which reads all
sheets from an .xlsx file to a list of data.frames.
}
\note{
This function requires openxlsx.
}
\examples{
x <- read_sheets("test.xlsx")
}
\author{
Kristian D. Olsen
}

