% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recode_entities.R
\name{recode_entities}
\alias{recode_entities}
\title{Recode entities based on open answers}
\usage{
recode_entities(survey, other = NULL, ...)
}
\arguments{
\item{survey}{A survey object.}

\item{other}{Optional: Name of the column containing open answers for mainentity.}

\item{...}{One or more named arguments.}
}
\description{
Small function to help in recoding the mainentity column based on open answers.
}
\examples{
x \%>\% recode_entites(other = "q1a", "Example" = "test")
}
\author{
Kristian D. Olsen
}

