% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporttool.R
\docType{package}
\name{reporttool}
\alias{reporttool}
\alias{reporttool-package}
\title{reporttool: From raw data to batch reports}
\description{
This is a package created to make working with and creating
reports from survey data easier. The main goal is to provide
a framework for turning raw data into batch reports in an efficient manner.
}
\details{
This is achieved by providing:

\itemize{
\item A predefined data structure, which facilitates most reporting jobs.
\item An intuitive pipeline for achieving the necessary structure.
\item Alternative i/o functions that make R work more seamlessly with every-day 
office tools.
\item Choice functions that make common operations less verbose.
\item Read-to-go templates for the supported formats (currently just PDF).
}
}

