% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{rmd_to_r}
\alias{rmd_to_r}
\title{Convert .Rmd to .R}
\usage{
rmd_to_r(rmd, encoding = "UTF-8", write = TRUE)
}
\arguments{
\item{rmd}{Path to a rmarkdown file.}

\item{encoding}{The encoding of both the input .Rmd file and the output.}
}
\value{
A .R file in the same directory and same name as the input .Rmd file.
}
\description{
This function converts a \code{.Rmd} to a \code{.R} file (similar to \code{knitr::purl}) 
by replacing chunk delimiters and putting \code{eval} chunk options inside \code{if}
statements. Also comments out (\code{##+}) content that is not inside chunks.
}
\note{
\code{UTF-8} is the recommended encoding for scripts and .Rmd files.
}
\examples{
rmd_to_r("Example report.Rmd", encoding = "latin1")
}
\author{
Kristian D. Olsen
}

