% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_association.R
\name{set_association}
\alias{set_association}
\title{Set latent associations for survey}
\usage{
set_association(survey, ..., common = FALSE)
}
\arguments{
\item{survey}{A survey object.}

\item{...}{The latent associations you would like to set. See examples.}

\item{common}{If set to \code{TRUE}, the function sets common associations
automatically. Typically, this will succesfully mark latents and mainentity.}
}
\description{
Use this function to specify latent associations and other variables
of special significance in the data. Beside latents, typical values are:
\code{mainentity subentity complaint oa_positive oa_negative
statement_categorical statement_metric} etc. You can also (auto) look for common
associations.
}
\note{
Use \code{none} to remove latent association. See examples.
}
\examples{
x <- survey(data.frame("test" = 1, stringsAsFactors = FALSE))
x <- x \%>\% add_mm() \%>\% set_association(mainentity = "q1", complaint = "q17")
x \%>\% add_entities() \%>\% add_weights()
x \%>\% set_association(none = "q1") # to remove
}
\author{
Kristian D. Olsen
}

