% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_colnames.R
\name{set_colnames}
\alias{set_colnames}
\title{Set latent associations for survey}
\usage{
set_colnames(survey, ..., nms = NULL)
}
\arguments{
\item{survey}{A survey object.}

\item{...}{Renaming of individual columns. Of the format \code{old_name = new_name}.}

\item{nms}{Optional character vector with the column names to use.}
}
\description{
Use this function to rename columns in the data in a survey. It makes sure
that the manifest stays updated as well.
}
\examples{
x \%>\% set_colnames(q1 = "mainentity")
}
\author{
Kristian D. Olsen
}

