% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_config.R
\name{set_config}
\alias{set_config}
\title{Set configuration for the survey}
\usage{
set_config(survey, ...)
}
\arguments{
\item{survey}{A survey object.}

\item{...}{The config you would like to set for the study.}
}
\description{
Use this function to specify config for the survey. Use it without arguments
to add default values.
}
\author{
Kristian D. Olsen
}

