% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_marketshares.R
\name{set_marketshares}
\alias{set_marketshares}
\title{Set marketshares for entities}
\usage{
set_marketshares(survey, ..., ms = NULL)
}
\arguments{
\item{survey}{A survey object.}

\item{...}{Marketshares for individual entities. Of the format \code{name = marketshare}.}

\item{ms}{Optional vector with the marketshares to use.}
}
\description{
Use this function to set marketshares for entities in Q1.
}
\examples{
x \%>\% set_marketshares("reporttool" = 1)
}
\author{
Kristian D. Olsen
}

