% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_translation.R
\name{set_translation}
\alias{set_translation}
\title{Set translations for the survey}
\usage{
set_translation(survey, ..., language = "norwegian")
}
\arguments{
\item{survey}{A survey object.}

\item{...}{The translations you would like to set.}

\item{language}{Default translations for a given language (currently Norwegian).}
}
\description{
Use this function to specify translations before generating reports. Use the
argument \code{language} to use a prespecified language (currently only Norwegian)
for the most common translations.
}
\author{
Kristian D. Olsen
}

