% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaries.R
\name{Summaries}
\alias{Summaries}
\alias{survey_info}
\alias{survey_table}
\title{Summaries}
\usage{
survey_info(srv, entity)

survey_table(srv, ..., entities = NULL, long_format = FALSE,
  questions = TRUE)
}
\description{
A few wrappers to make a common tasks less verbose.
}
\section{List of summaries}{


\describe{

   \item{\code{survey_info}}{Returns a list of information about the survey
   and entity, including respondents, dates of first and last response etc.}

   \item{\code{survey_table}}{Creates a table based on survey data, in either
   wide or long format. The function filters missing_percentage based on cutoff,
   and includes a weighted average as the last row in the table. It also
   replaces columnnames with the question text and/or the specified translations.
   Note: The function suppresses warnings on left_join and filter.}

}
}
\examples{
survey_info(srv)
}
\author{
Kristian D. Olsen
}

