% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survey.R
\name{survey}
\alias{as.survey}
\alias{as.survey.default}
\alias{as.survey.survey}
\alias{is.survey}
\alias{print.survey}
\alias{survey}
\title{Create a survey object}
\usage{
survey(x)

is.survey(x)

as.survey(x)

\method{as.survey}{default}(x)

\method{as.survey}{survey}(x)

\method{print}{survey}(x, width = getOption("width"))
}
\arguments{
\item{x}{A list or a data.frame}
}
\description{
First step in creating a \code{survey} that can be used with \code{generate_report}.
The function accepts either a \code{list} or a \code{data.frame} as input, and
will only retain information following the structure and naming convention.
}
\note{
When using \code{\link{write_data}} with a \code{survey} object, all names will
be converted to their longer versions (alias). E.g. "df" becomes "data" and so forth.
Conversely, when converting a \code{list} to \code{survey} with \code{survey} or
\code{as.survey}, long names are converted to their short version. This is
done for brevity when coding.
}
\section{Structure/naming convention for \code{list}}{


\describe{

   \item{\code{df}}{The dataset from the current study, either in it's processed or raw
   format. (Alias: \code{data}).}

   \item{\code{cd}}{Optional: A dataset to compare or contrast the current study to.
   Typically used when comparing a smaller (web) study to the national study.
   (Alias: \code{contrast data}).}

   \item{\code{hd}}{Optional: Historical data used to compare the present study against.
   (Alias: \code{historical data}).}

   \item{\code{ents}}{Data regarding the entities (typically Q1), and their
   respecitve marketshares etc. (Alias: \code{entities}).}

   \item{\code{mm}}{The measurement model (questionnaire) and additional information
   describing the variables in the data. (Alias: \code{measurement model}).}

   \item{\code{cfg}}{Config for the study and \code{generate_report}, including
   translations/printed names for latent variables. (Alias: \code{config}).}
}

A structure and naming convention is also required for the measurement model,
entities and config. See \code{\link{add_model}}, \code{\link{add_entities}} and
\code{\link{add_config}} for information.
}
\examples{
x <- survey(data.frame("test" = 1, stringsAsFactors = FALSE))
}
\author{
Kristian D. Olsen
}

