% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaries.R
\name{survey_table}
\alias{survey_table}
\title{Create tables from surveys}
\usage{
survey_table(srv, ..., drop = FALSE, wide = TRUE, weighted = TRUE,
  questions = TRUE, contrast = TRUE)
}
\arguments{
\item{srv}{A survey object.}

\item{...}{Columns to summarise. All columns must be of the same type, and not
text columns. If they are several factor variables, they must have the same levels.}

\item{drop}{Default is \code{FALSE}, which means unused factor levels will be
kept in the data and filled with \code{0} for factors, or \code{NA} for numeric.}

\item{wide}{If this is \code{TRUE} (the default), the output will be a wide
\code{data.frame}.}

\item{weighted}{When \code{TRUE}, the average will be weighted.}

\item{questions}{When \code{TRUE}, the question text specified in the measurement
model will be included in the table (if they are not empty strings).}

\item{contrast}{Set to false if a contrast exist but you want to use the study
average instead.}
}
\description{
Function for creating summary tables of factors and numeric columns in a survey.
This function will always group by the mainentity in a survey, and always add
the average for the study. You can also group the survey by other variables
before passing it to \code{survey_table} to get scores/proportions of an
arbitrary variable, grouped gender in addition to company for instance.
}
\examples{
x \%>\% group_by(q7_service) \%>\% survey_table(image:loyal)
}
\author{
Kristian D. Olsen
}

