% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topline.R
\name{topline}
\alias{topline}
\title{Create a topline}
\usage{
topline(df, mainentity = "q1", me_open = "q1_open", scores = c("q3", "q6",
  "q16"))
}
\arguments{
\item{df}{The data.frame containing the study data}

\item{mainentity}{Name of the mainentity (q1) column in the data}

\item{me_open}{The text column which contains the open response for 'other' in
mainentity.}

\item{scores}{A vector with columnnames for variables you would like a mean from.}
}
\value{
A list
}
\description{
Simple function to get a count and mean per entity in an ongoing study, as well
as open answers for respondents who have answered 'other'.
}
\examples{
x <- topline(df)
}
\author{
Kristian D. Olsen
}

