% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaults.R, R/utilities.R
\name{get_default}
\alias{Utilities}
\alias{clean_score}
\alias{get_default}
\alias{intranet_link}
\alias{ordered_replace}
\alias{recode}
\alias{rescale_score}
\title{Utilities}
\usage{
get_default(string)

recode(x, ..., by = x, drop = TRUE)

clean_score(var)

rescale_score(var)

ordered_replace(x, match_by, replacement = NULL)

intranet_link(https)
}
\description{
A few wrappers to make a common tasks less verbose.
}
\section{List of utilities}{


\describe{

   \item{\code{recode}}{Recode variables without nested ifelse statements. Uses
   \code{\%in\%} to create a logical vector when recoding. I.e. you can recode one
   vector based on matches in another vector of the same length. For factors,
   the function only allows recoding to an existing level, but you can set
   \code{drop = TRUE} to drop levels that you are recoding (away) from.}

   \item{\code{intranet_link}}{Converts any http(s) link with a .se domain
   to a link for a network drive (sharepoint) on windows.}

   \item{\code{clean_score}}{Takes vectors representing likert scales and
   cleans text descriptions. E.g. "10 Very happy" becomes "10", and the value
   is converted to \code{numeric} without warning.}

   \item{\code{rescale_score}}{Takes vectors representing 10-point likert
   scales and transforms them to 100-point scales (\code{numeric}). (x-1)*(100/9)}

   \item{\code{ordered_replace}}{Replace \code{x} with \code{replacement} where \code{x}
   matches \code{match_by}. Matches and replacements retain the original order of
   \code{x}.}

   \item{\code{get_default}}{Get default settings from the reporttool package.
   \code{x} is a search-string which is not case sensitive.}

   \item{\code{set_missing}}{Takes a \code{data.frame} and cleans
   the default missing value strings. See \code{defaults.R} for these
   strings.}

}
}
\examples{
get_default("palette")
}
\author{
Kristian D. Olsen
}

