% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_output.R
\name{write_data}
\alias{write_data}
\title{Write common file formats}
\usage{
write_data(x, file = NULL, encoding = "latin1")
}
\arguments{
\item{x}{The data to be written. Can be a data.frame or a list (of df's).}

\item{file}{Path and extension for saving the data.}

\item{encoding}{The encoding to use for txt and csv-files.}
}
\description{
A simple wrapper for writing common data formats. The specific format is
specified by the extension given in the filename. If no filename is given,
data will be written to output.xlsx in the current working directory.
}
\note{
Writing .xlsx requires that the openxlsx package is installed.
}
\examples{
write_data(x, file = "test.xlsx")
}
\author{
Kristian D. Olsen
}

