% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{write_questionnaire}
\alias{write_questionnaire}
\title{Write questionnaire}
\usage{
write_questionnaire(quest, file, study = "Banking", segment = "B2C",
  entity = NULL)
}
\arguments{
\item{quest}{The questionnaire (r object).}

\item{file}{The name of the xlsx file you would like to write to.}

\item{study}{Name of the study to write.}

\item{segment}{The segment being studied.}

\item{entity}{Optional: Replace {XX} with something else.}
}
\description{
This function writes a questionnaire as extracted by get_questionnaire to
a specified xlsx-file (with styling). Only writes the latent association,
manifest name, possible values and the question text itself.
}
\examples{
q <- read_data("master questionnaire.xlsx", sheet = "questionnaires")
write_questionnaire(q, "Banking B2C.xlsx", study = "Banking", segment = "B2C")
}
\author{
Kristian D. Olsen
}

