% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{write_sharepoint}
\alias{write_sharepoint}
\title{Write input files}
\usage{
write_sharepoint(srv, file)
}
\arguments{
\item{srv}{A survey object.}

\item{file}{The directory on sharepoint where you would like to write the input
files.}
}
\description{
This function creates the necessary input files for model estimation, in the
specified directory.
}
\examples{
write_sharepoint("https://the.intranet.se/EPSI/example")
}
\author{
Kristian D. Olsen
}

