% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_missing}
\alias{clean_missing}
\alias{clean_score}
\alias{ordered_replace}
\alias{rescale_score}
\title{Utilities}
\usage{
clean_missing(df, na.strings = reporttool$missing_values)

clean_score(var)

rescale_score(var)

ordered_replace(x, by, y)
}
\arguments{
\item{df}{A \code{data.frame}}

\item{na.strings}{Optional: Provide a list of strings to convert to NA.}

\item{var}{A numeric or character vector.}
}
\value{
\code{clean_missing} returns a \code{data.frame}, \code{clean_score}
returns a vector of the same type (typically character) and \code{rescale_score}
returns a numeric vector.
}
\description{
A few wrappers to make a common tasks less verbose.
}
\section{List of utilities}{


\describe{
   \item{\code{ordered_replace}}{Replace \code{x} with \code{y} where \code{x}
   matches \code{by}. Matches and replacements retain the original order of
   \code{x}.}

   \item{\code{clean_missing}}{Takes a \code{data.frame} and cleans
   the default missing value strings. See \code{defaults.R} for these
   strings.}

   \item{\code{clean_score}}{Takes vectors representing likert scales and
   cleans text descriptions. E.g. "10 Very happy" becomes "10".}

   \item{\code{rescale_score}}{Takes vectors representing 10-point likert
   scales and transforms them to 100-point scales. (x-1)*(100/9)}
}
}
\examples{
df <- clean_missing(df)
df$Q3 <- clean_score(df$Q3)
df$Q3 <- rescale_score(df$Q3)
}
\author{
Kristian D. Olsen
}

