% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_ppt.R
\name{generate_ppt}
\alias{generate_ppt}
\title{Powerpoint template}
\usage{
generate_ppt(entity, dir, envir)
}
\arguments{
\item{entity}{Name of the entity.}

\item{dir}{Base directory (from .Rmd template)}

\item{envir}{The environment in which to evaluate the rmarkdown code.}
}
\description{
This function uses \code{ReporteRs} and \code{evaluate} to generate powerpoint
slides based on the same rmarkdown syntax that creates the pdf reports. I.e.
it is a dumbed down version of normal rmarkdown, where only \code{#} and \code{##}
are understood as being section titles and new frames/subtitles respectively.
}
\note{
This function is used by \code{generate_report}.
}
\author{
Kristian D. Olsen
}

