% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate report}
\usage{
generate_report(srv, report = NULL, entity = NULL, type = "pdf")
}
\arguments{
\item{srv}{A survey object containing the data that is being reported.}

\item{report}{Path to the report template (.Rmd) for the specific study.}

\item{entity}{Optional: Name of a specific entity to create a report for.}
}
\value{
Nothing. A .Rmd file and desired output is created for each entity in
the data, or as specified.
}
\description{
Generate one or more reports using the templates in this package.
}
\examples{
generate_report(x, "/Internal/Internal_report_2014.Rmd", entity=c("EPSI", "SKI"))
}
\author{
Kristian D. Olsen
}

