% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{get_questionnaire}
\alias{get_questionnaire}
\title{Get a questionnaire with exchanged placeholders}
\usage{
get_questionnaire(file, industry = "Energy", study = "District Heating",
  entity = NULL)
}
\arguments{
\item{file}{Path to the master questionnaire}

\item{industry}{Specify which industry the study you would like to extract
belongs to}

\item{study}{The study to extract and replace values for.}

\item{entity}{Optional: Replace {XX} in the questionnaire with whatever you
specify.}
}
\description{
A function for internal use. Reads the master questionnaire and exchanges
placeholders in the text with their respective dictionary entry.
}
\examples{
x <- get_questionnaire("masterquest.xlsx", "Energy", "Electricity", entity = "[S2]")
}
\author{
Kristian D. Olsen
}

