% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare survey data}
\usage{
prepare_data(survey, type = NULL)
}
\arguments{
\item{input}{Either a path to a file, or a list containing the raw data and/or
measurement model/entities.}

\item{latents}{Specify either 'pls' or 'mean' to calculate latents based on
the latent specification in the measurement model. Requires that latents are in
fact specified.}

\item{impute}{Specify whether missing values should be imputed for observations
that are retained.}

\item{cutoff}{The missing-values cutoff; defaults to .3 (30%). Any observations
with a missing percent above this threshold are excluded when imputing missing
values and calculating latents.}
}
\value{
A list containing the processed data, measurement model and entities.
}
\description{
This function expects an input file containing raw data. Ideally the input also
includes a measurement model and entities w/marketshares, but it attempts to
create these based on the data (to help achieve the desired structure) if they
are not found.
}
\examples{
prepared <- prepare_data("test.xlsx", latents = "pls", impute = TRUE, cutoff = .3)
}
\author{
Kristian D. Olsen
}

