% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare survey data}
\usage{
prepare_data(survey, type = NULL)
}
\arguments{
\item{survey}{A survey object.}

\item{type}{Either \code{mean} or \code{pls}. Default is \code{NULL}, in which
case \code{mean} is used.}
}
\value{
Returns the survey with EM-variables and latent scores added using the
specified method.
}
\description{
This function does the imputation/PLS-PM (or simple mean) for the survey data.
In addition, it calculates missing percentage and tallies the valid observations
for each entity.
}
\author{
Kristian D. Olsen
}

