% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare survey data}
\usage{
prepare_data(srv, type = "mean", cutoff = 0.3)
}
\arguments{
\item{srv}{A survey object.}

\item{type}{The approach to use for latents. Either \code{mean}, \code{pls},
or \code{none} (no latents). If \code{pls} is used, only \code{coderesp} and
\code{percent_missing} are added to the data. I.e. it is a preparation for the
PLS-wizard.}
}
\value{
Returns the survey with EM-variables and latent scores added using the
specified method.
}
\description{
This function does the imputation/PLS-PM (or simple mean) for the survey data.
In addition, it calculates missing percentage and tallies the valid observations
for each entity.
}
\author{
Kristian D. Olsen
}

