% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_data}
\alias{read_data}
\title{Read common data formats}
\usage{
read_data(file, encoding = "latin1")
}
\arguments{
\item{file}{Path to a csv, csv2 or xlsx file.}

\item{encoding}{The encoding to use for txt and csv-files.}
}
\value{
A data.frame. If more than one sheet is read from a xlsx file a
list is returned instead.
}
\description{
A simple wrapper for reading data. Currently supports
csv, csv2 and xlsx. (TODO: .txt and study-directories).
}
\note{
Reading .xlsx requires that the openxlsx package is installed.
}
\examples{
x <- read_data("test.xlsx")
}
\author{
Kristian D. Olsen
}

