% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_translation.R
\name{set_translation}
\alias{get_translation}
\alias{set_translation}
\title{Set translations for the survey}
\usage{
set_translation(srv, ..., language = "norwegian")

get_translation(srv, translation)
}
\arguments{
\item{srv}{A survey object.}

\item{...}{The translations you would like to set.}

\item{language}{Default translations for a given language (currently Norwegian).}
}
\description{
Use this function to specify translations before generating reports. Use the
argument \code{language} to use a prespecified language (currently only Norwegian) 
for the most common translations.
}
\author{
Kristian D. Olsen
}

