% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survey_table.R
\name{survey_table}
\alias{survey_table}
\alias{survey_table_}
\title{Create tables from surveys}
\usage{
survey_table(srv, ..., wide = TRUE, weight = TRUE, question = TRUE,
  filter_missing = TRUE, filter_response = TRUE, contrast = TRUE)

survey_table_(srv, dots, wide = TRUE, weight = TRUE, question = TRUE,
  filter_missing = TRUE, filter_response = TRUE, contrast = TRUE)
}
\arguments{
\item{srv}{A survey object.}

\item{...}{Columns to summarise. Accepts either numeric or factors, and a
warning will be issued if it encounters text etc. When creating a table for
factors, they must all have identical factor levels.}

\item{wide}{If this is \code{TRUE} (the default), the output will be in a wide format.}

\item{weight}{When \code{TRUE}, the average will be weighted (assumes weights are in column 'w').}

\item{question}{When \code{TRUE}, the question text specified in the measurement
model will be included in the table (provided they are not empty strings).}

\item{filter_missing}{Set to \code{FALSE} to NOT remove observations with \code{percent_missing}
above the cutoff that is set in config.}

\item{filter_response}{When set to \code{TRUE} (default), the function will
filter NA values in the response variables (i.e. \code{...}) before counting
observations and/or calculating mean or proportions}

\item{contrast}{Set to \code{FALSE} if a contrast exist but you want to use the study
average instead.}
}
\description{
Function for creating summary tables of factors and numeric columns in a survey.
You can also group the survey by other variables before passing it to \code{survey_table}
to get scores/proportions by one or more variables.
}
\details{
The function does the following:

\describe{

\item{\code{average}}{Produces an average for the tables, which is always the
   average for the first group (as set by \code{group_by}),
   and it is appropriately grouped for the remaining groups.}

\item{\code{weight}}{Unless otherwise specified, this function will always
   weight the results for the average. This also means that the function requires
   that the weight column (\code{w}) is present in the data.}

\item{\code{filter}}{Filters missing values \code{NA}
   for all groups and the variables that go into the table, as well as
   \code{percent_missing} which exceeds the cutoff in the surveys config. Setting
   \code{filter_missing} to \code{FALSE} overrides the filtering of \code{percent_missing}.}

\item{\code{count}}{For both numeric and factor variables, this function always
   provides a count for the number of valid observations (after filtering) in each group.}

\item{\code{missing}}{Turns implicit missing values into
   explicit missing values. When this is the case, counts will be 0 and
   the aggregated values will be \code{NA} (unless it is a factor and one of the
   other proportions are not NA, in this case, the proportions are set to 0 instead.)}

\item{\code{spread}}{When \code{wide} is set to \code{TRUE}, the function
   spreads the results to a wide format. For numeric, the variables are put in
   separate columns. For factor variables, the proportions are
   spread by their respective levels (e.g. "Yes", "No" etc become columns.))}

}
}
\examples{
x \%>\% group_by(q7_service) \%>\% survey_table(image:loyal)
}
\author{
Kristian D. Olsen
}

