% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_output.R
\name{to_clipboard}
\alias{to_clipboard}
\title{Write to Windows clipboard}
\usage{
to_clipboard(df, names = TRUE, encoding = "")
}
\arguments{
\item{df}{The data to write.}

\item{names}{Should columnnames be included?}

\item{encoding}{The encoding to use when writing.}
}
\description{
Wrapper for writing to windows clipboard with the most-used defaults for a
scandinavian locale.
}
\note{
This function only works on Windows, and the data-size cannot exceeed 128kb.
}
\examples{
x \%>\% to_clipboard()
}
\author{
Kristian D. Olsen
}

