% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_output.R
\name{to_sheet}
\alias{to_sheet}
\title{Write data to sheet (in an openxlsx workbook)}
\usage{
to_sheet(df, wb, sheet = "analysis", row = 1L, append = TRUE)
}
\arguments{
\item{df}{The data to write.}

\item{wb}{A loaded openxlsx workbook (use openxlsx::loadWorkbook)}

\item{sheet}{Name of the sheet to write to, will be created if it does not exist.}

\item{row}{Optional: Also specify the startingrow for writing data.}

\item{append}{Whether or not the function should append data or clean the
sheet before writing.}
}
\value{
A list containing data.frames matching the sheets in the .xlsx file.
If only one sheet is read, the function returns a data.frame instead.
}
\description{
This function (a very thin wrapper for openxlsx functions) that writes/appends
a data.frame to the specified sheet in a loaded openxlsx workbook.
}
\note{
This function requires openxlsx.
}
\examples{
wb <- openxlsx::loadWorkbook("test.xlsx")
x \%>\% to_sheet(wb, sheet = "test", append = FALSE)
openxlsx::saveWorkbook(wb, "test.xlsx", overwrite = TRUE)
}
\author{
Kristian D. Olsen
}

