% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaults.R, R/summaries.R, R/utilities.R
\name{get_default}
\alias{Summaries}
\alias{Utilities}
\alias{clean_score}
\alias{get_default}
\alias{info_list}
\alias{intranet_link}
\alias{ordered_replace}
\alias{recode}
\alias{rescale_score}
\alias{set_missing}
\title{Summaries}
\usage{
get_default(string)

info_list(srv, entity)

recode(x, ..., by = x)

set_missing(df, na_strings = get_default("na_strings"))

clean_score(var)

rescale_score(var)

ordered_replace(x, match_by, replacement = NULL)

intranet_link(https)
}
\description{
A few wrappers to make a common tasks less verbose.

A few wrappers to make a common tasks less verbose.
}
\section{List of summaries}{


\describe{

   \item{\code{recode}}{Recode variables.}

}
}

\section{List of utilities}{


\describe{

   \item{\code{recode}}{Recode variables.}

   \item{\code{intranet_link}}{Converts any http(s) link with a .se domain
   to a link for a network drive (sharepoint) on windows.}

   \item{\code{clean_score}}{Takes vectors representing likert scales and
   cleans text descriptions. E.g. "10 Very happy" becomes "10", and the value
   is converted to \code{numeric} without warning.}

   \item{\code{rescale_score}}{Takes vectors representing 10-point likert
   scales and transforms them to 100-point scales (\code{numeric}). (x-1)*(100/9)}

   \item{\code{ordered_replace}}{Replace \code{x} with \code{replacement} where \code{x}
   matches \code{match_by}. Matches and replacements retain the original order of
   \code{x}.}

   \item{\code{get_default}}{Get default settings from the reporttool package.
   \code{x} is a search-string which is not case sensitive.}

   \item{\code{set_missing}}{Takes a \code{data.frame} and cleans
   the default missing value strings. See \code{defaults.R} for these
   strings.}

}
}
\examples{
get_default("palette")
get_default("palette")
}
\author{
Kristian D. Olsen

Kristian D. Olsen
}

