% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaults.R, R/utilities.R
\name{rt_defaults}
\alias{Utilities}
\alias{clean_score}
\alias{intranet_link}
\alias{lowercase_names}
\alias{ordered_replace}
\alias{rescale_score}
\alias{rt_defaults}
\alias{set_missing}
\title{Utilities}
\usage{
rt_defaults(x)

set_missing(df, na.strings = cfg$missing_values)

clean_score(var)

rescale_score(var)

ordered_replace(x, match_by, replacement = NULL)

intranet_link(https)

lowercase_names(x)
}
\description{
A few wrappers to make a common tasks less verbose.
}
\section{List of utilities}{


\describe{
   \item{\code{rt_defaults}}{Get default settings from the reporttool package.
   \code{x} is a search-string which is not case sensitive.}

   \item{\code{intranet_link}}{Converts any http(s) link with a .se domain
   to a link for a network drive (sharepoint) on windows.}

   \item{\code{ordered_replace}}{Replace \code{x} with \code{replacement} where \code{x}
   matches \code{match_by}. Matches and replacements retain the original order of
   \code{x}.}

   \item{\code{set_missing}}{Takes a \code{data.frame} and cleans
   the default missing value strings. See \code{defaults.R} for these
   strings.}

   \item{\code{clean_score}}{Takes vectors representing likert scales and
   cleans text descriptions. E.g. "10 Very happy" becomes "10".}

   \item{\code{rescale_score}}{Takes vectors representing 10-point likert
   scales and transforms them to 100-point scales. (x-1)*(100/9)}

   \item{\code{lowercase_names}}{Used to lowercase columnnames in lists
   (instead of an anonyomous function.)}
}
}
\examples{
df <- set_missing(df)
df$Q3 <- clean_score(df$Q3)
df$Q3 <- rescale_score(df$Q3)
ordered_replace("measurement model", with(cfg$sheet_names, setNames(long, short)))
lst <- lapply(lst, lowercase_names)
rt_defaults("latent")
}
\author{
Kristian D. Olsen
}

