% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{Utilities}
\alias{Utilities}
\alias{clean_missing}
\alias{clean_score}
\alias{get_sheet_names}
\alias{ordered_replace}
\alias{rescale_score}
\title{Utilities}
\usage{
clean_missing(df, na.strings = reporttool$missing_values)

clean_score(var)

rescale_score(var)

ordered_replace(x, by, y)

get_sheet_names(file)
}
\description{
A few wrappers to make a common tasks less verbose.
}
\section{List of utilities}{


\describe{
   \item{\code{get_sheet_names}}{Takes the path to a .xlsx file and returns
   the sheetnames in the file if any sheets exist. Quicker than using
   openxlsx::loadWorkbook() and names() when you do not need to preserve
   styling/formatting.}

   \item{\code{ordered_replace}}{Replace \code{x} with \code{y} where \code{x}
   matches \code{by}. Matches and replacements retain the original order of
   \code{x}.}

   \item{\code{clean_missing}}{Takes a \code{data.frame} and cleans
   the default missing value strings. See \code{defaults.R} for these
   strings.}

   \item{\code{clean_score}}{Takes vectors representing likert scales and
   cleans text descriptions. E.g. "10 Very happy" becomes "10".}

   \item{\code{rescale_score}}{Takes vectors representing 10-point likert
   scales and transforms them to 100-point scales. (x-1)*(100/9)}
}
}
\examples{
df <- clean_missing(df)
df$Q3 <- clean_score(df$Q3)
df$Q3 <- rescale_score(df$Q3)
}
\author{
Kristian D. Olsen
}

