% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{write_questionnaire}
\alias{write_questionnaire}
\title{Write questionnaire}
\usage{
write_questionnaire(quest, file, study = "District heating")
}
\arguments{
\item{quest}{A questionnaire as returned by get_questionnaire.}

\item{file}{The name of the xlsx file you would like to write to.}

\item{study}{Name of the study. This is used as the sheet-name when writing
the questionnaire.}
}
\description{
This function writes a questionnaire as extracted by get_questionnaire to
a specified xlsx-file (with styling). Only writes the latent association,
manifest name, possible values and the question text itself.
}
\examples{
studies <- c("Mobile", "Broadband", "DTV")
all_questionnaires <- lapply(studies, get_questionnaire, file = masterquest.xlsx, industry = "Telecom")
lapply(all_questionnaires, write_questionnaire, file = "ICT Questionnaires.xlsx", study = studies)
}
\author{
Kristian D. Olsen
}

