% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/em_functions.R
\name{write_sharepoint}
\alias{write_sharepoint}
\title{Extract from a study directory on sharepoint}
\usage{
write_sharepoint(lst, dir)
}
\arguments{
\item{lst}{A survey_data object.}

\item{dir}{The directory on sharepoint where you would like to write the input
files.}
}
\value{
A list containing the EM-data, entities (observations, marketshare),
and a measurement model (latent association, question text, values etc.)
}
\description{
This function extracts whatever data is available in our standard file
structure and tries to structure it like a input.xlsx file.
}
\examples{
x <- read_em("https://the.intranet.se/EPSI/example")
}
\author{
Kristian D. Olsen
}

