% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io_sharepoint.R
\name{write_sharepoint}
\alias{write_sharepoint}
\title{Write input files}
\usage{
write_sharepoint(lst, dir)
}
\arguments{
\item{lst}{A survey_data object.}

\item{dir}{The directory on sharepoint where you would like to write the input
files.}
}
\value{
A list containing the EM-data, entities (observations, marketshare),
and a measurement model (latent association, question text, values etc.)
}
\description{
This function creates the necessary input files for model estimation, in the
specified directory.
}
\examples{
x <- write_sharepoint("https://the.intranet.se/EPSI/example")
}
\author{
Kristian D. Olsen
}

